lab6/requirement.txt
contourpy==1.3.2
cycler==0.12.1
filelock==3.16.1
fonttools==4.57.0
fsspec==2024.10.0
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
packaging==25.0
pillow==11.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-triton
setuptools==70.2.0
six==1.17.0
sympy==1.13.3
torch
torchaudio
torchvision
tqdm==4.67.1
typing_extensions==4.12.2
